requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blobfile==3.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
DALL-E @ git+https://github.com/openai/DALL-E.git@5be4b236bc3ade6943662354117a0e83752cc322
debugpy==1.8.14
decorator==5.2.1
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.19.1
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
grpcio==1.75.0
hf-xet==1.1.10
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
iniconfig==2.1.0
ipykernel==6.29.5
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1748711175/work
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lightning-utilities==0.15.2
lxml==6.0.1
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.4
mypy==1.18.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
omegaconf==2.3.0
opencv-python-headless==4.12.0.88
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodomex==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pyDeprecate==0.3.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scipy==1.15.3
simsimd==6.5.3
six==1.17.0
stack-data==0.6.3
stringzilla==4.0.14
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
test_tube==0.7.5
tokenizers==0.22.0
tomli==2.2.1
torch==2.8.0
torchaudio==2.8.0
torchmetrics==0.4.0
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wget==3.2
yarl==1.20.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
